working setup for buffering param changes
