+++ frontend/src/app/services/param.service.ts
@@ -1,30 +1,22 @@
 import { Injectable } from '@angular/core';
-import { ActivatedRoute, Params, Router } from '@angular/router';
-import { BehaviorSubject, Subscription } from 'rxjs';
+import { Params } from '@angular/router';
+import { BehaviorSubject, Observable } from 'rxjs';
 import { bufferTime } from 'rxjs/operators';
 
 @Injectable({
   providedIn: 'root'
 })
 export class ParamService {
+  /**
+   * Provides a subject for query parameter changes
+   * Buffers updates in the bufferedParams Observable, subscribed to by ParamDirective
+   */
 
+  public bufferedParams = new Observable<Params[]>();
   private currentParams =  new BehaviorSubject<Params>({});
-  private buffer: Subscription;
 
-  constructor(private router: Router) {
-    this.buffer = this.currentParams.pipe(bufferTime(100)).subscribe( params => {
-      if (params.length) {
-        const newParams = Object.assign({}, ...params.map(f => f));
-        console.log(newParams);
-        this.router.navigate(
-          [this.router.url],
-          {
-            queryParams: newParams,
-            queryParamsHandling: 'merge',
-          }
-        );
-      }
-    });
+  constructor() {
+    this.bufferedParams = this.currentParams.pipe(bufferTime(500));
   }
 
   setParams(params: Params) {
